+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## [0.1.0]
 
 * First version
-* WebSocket communication for private topics and klines
+* WebSocket and HTTPS (REST) communication